requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-binance==1.0.19
openai>=1.3.9
gunicorn==22.0.0
python-dotenv==1.0.1
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
eventlet==0.33.3
cryptography==42.0.5
pycryptodome==3.20.0
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.9
requests==2.31.0
APScheduler==3.10.4
